+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 * Fixes main thread violation on Android
 * Fixes an issue with request and response headers parameters type causing network requests not getting logged on iOS
 * Uses pigeon for internal communication between Flutter and the host platform
+* Uses return values instead of callbacks in the following APIs:
+  1. Replies.getUnreadRepliesCount
+  2. Replies.hasChats
+  3. Surveys.hasRespondedToSurvey
+  4. Surveys.getAvailableSurveys
 
 ## 11.3.0 (2022-09-30)